+++ ios/Classes/TiMapView.m
@@ -487,7 +487,7 @@
 		} else if ([[NSBundle mainBundle] objectForInfoDictionaryKey:@"NSLocationWhenInUseUsageDescription"]){
 			[locationManager requestWhenInUseAuthorization];
 		} else {
-			NSLog(@"[ERROR] The keys NSLocationAlwaysUsageDescription or NSLocationWhenInUseUsageDescription are not defined in your tiapp.xml.  Starting with iOS8 this is required.");
+			NSLog(@"[ERROR] The keys NSLocationAlwaysUsageDescription or NSLocationWhenInUseUsageDescription are not defined in your tiapp.xml. Starting with iOS 8 this is required.");
 		}
 		// Create the map
 		[self map];
@@ -727,17 +727,28 @@
     [TiMapModule logAddedIniOS7Warning:@"showsPointsOfInterest"];
 }
 
+-(void)setShowsCompass_:(id)value
+{
+    [TiMapModule logAddedIniOS7Warning:@"showsCompass"];
+}
+
+-(void)setShowsScale_:(id)value
+{
+    [TiMapModule logAddedIniOS7Warning:@"showsScale"];
+}
+
+-(void)setShowsTraffic_:(id)value
+{
+    [TiMapModule logAddedIniOS7Warning:@"showsTraffic"];
+}
+
+
 #pragma mark Utils
 // Using these utility functions allows us to override them for different versions of iOS
 
 -(void)addOverlay:(MKPolyline*)polyline level:(MKOverlayLevel)level
 {
     [map addOverlay:polyline];
-}
-
--(BOOL)isIOS9OrGreater
-{
-    return [UIImage instancesRespondToSelector:@selector(flipsForRightToLeftLayoutDirection)];
 }
 
 #pragma mark Delegates
@@ -764,6 +775,17 @@
     return (MKOverlayView *)CFDictionaryGetValue(mapObjects2View, overlay);
 }
 
+-(void)mapView:(MKMapView *)mapView regionWillChangeAnimated:(BOOL)animated
+{
+    if (ignoreRegionChanged) {
+        return;
+    }
+    
+    if ([[self proxy] _hasListeners:@"regionwillchange"]) {
+        [self fireEvent:@"regionwillchange" withRegion:region animated:animate];
+    }
+}
+
 - (void)mapView:(MKMapView *)mapView regionDidChangeAnimated:(BOOL)animated
 {
     if (ignoreRegionChanged) {
@@ -771,26 +793,14 @@
     }
     region = [mapView region];
     [self.proxy replaceValue:[self dictionaryFromRegion] forKey:@"region" notification:NO];
-	if ([self.proxy _hasListeners:@"regionChanged"])
-	{	//TODO: Deprecate old event
-		NSDictionary * props = [NSDictionary dictionaryWithObjectsAndKeys:
-								@"regionChanged",@"type",
-								[NSNumber numberWithDouble:region.center.latitude],@"latitude",
-								[NSNumber numberWithDouble:region.center.longitude],@"longitude",
-								[NSNumber numberWithDouble:region.span.latitudeDelta],@"latitudeDelta",
-								[NSNumber numberWithDouble:region.span.longitudeDelta],@"longitudeDelta",nil];
-		[self.proxy fireEvent:@"regionChanged" withObject:props];
+	
+    if ([self.proxy _hasListeners:@"regionChanged"]) {
+        NSLog(@"[WARN] The 'regionChanged' event is deprecated, use 'regionchanged' instead.");
+        [self fireEvent:@"regionChanged" withRegion:[mapView region] animated:animate];
 	}
-	if ([self.proxy _hasListeners:@"regionchanged"])
-	{
-		NSDictionary * props = [NSDictionary dictionaryWithObjectsAndKeys:
-								@"regionchanged",@"type",
-								[NSNumber numberWithDouble:region.center.latitude],@"latitude",
-								[NSNumber numberWithDouble:region.center.longitude],@"longitude",
-								[NSNumber numberWithDouble:region.span.latitudeDelta],@"latitudeDelta",
-								[NSNumber numberWithDouble:region.span.longitudeDelta],@"longitudeDelta",
-								NUMBOOL(animated),@"animated",nil];
-		[self.proxy fireEvent:@"regionchanged" withObject:props];
+    
+	if ([self.proxy _hasListeners:@"regionchanged"]) {
+        [self fireEvent:@"regionchanged" withRegion:[mapView region] animated:animate];
 	}
 }
 
@@ -968,18 +978,7 @@
         else {
             MKPinAnnotationView *pinview = (MKPinAnnotationView*)annView;
             
-            if([self isIOS9OrGreater] == YES) {
-#if __IPHONE_9_0
-                pinview.pinTintColor = [ann pinColor];
-#endif
-            } else {
-                
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-                pinview.pinColor = [ann pinColor];
-#pragma GCC diagnostic pop
-            }
-            
+            pinview.pinColor = [ann pincolor];
             pinview.animatesDrop = [ann animatesDrop] && ![(TiMapAnnotationProxy *)annotation placed];
             annView.calloutOffset = CGPointMake(-8, 0);
         }
@@ -1111,7 +1110,7 @@
         CGPoint polygonViewPoint = [polygonRenderer pointForMapPoint:point];
         BOOL inPolygon = CGPathContainsPoint(polygonRenderer.path, NULL, polygonViewPoint, NO);
         if (inPolygon) {
-            [self fireShapeClickEvent:proxy point:point sourceType:VIEW_TYPE_POLYGON];
+            [self fireShapeClickEvent:proxy point:point sourceType:@"polygon"];
         }
     }
 }
@@ -1125,7 +1124,7 @@
         CGPoint circleViewPoint = [circRenderer pointForMapPoint:point];
         BOOL inCircle = CGPathContainsPoint(circRenderer.path, NULL, circleViewPoint, NO);
         if (inCircle) {
-            [self fireShapeClickEvent:circle point:point sourceType:VIEW_TYPE_CIRCLE];
+            [self fireShapeClickEvent:circle point:point sourceType:@"circle"];
         }
     }
 }
@@ -1139,9 +1138,22 @@
         CGPoint polylineViewPoint = [polylineRenderer pointForMapPoint:point];
         BOOL onPolyline = CGPathContainsPoint(polylineRenderer.path, NULL, polylineViewPoint, NO);
         if (onPolyline) {
-            [self fireShapeClickEvent:proxy point:point sourceType:VIEW_TYPE_POLYLINE];
+            [self fireShapeClickEvent:proxy point:point sourceType:@"polyline"];
         }
     }
+}
+
+- (void)fireEvent:(NSString*)event withRegion:(MKCoordinateRegion)_region animated:(BOOL)animated
+{
+    NSDictionary *object = [NSDictionary dictionaryWithObjectsAndKeys:
+     event,@"type",
+     [NSNumber numberWithDouble:_region.center.latitude],@"latitude",
+     [NSNumber numberWithDouble:_region.center.longitude],@"longitude",
+     [NSNumber numberWithDouble:_region.span.latitudeDelta],@"latitudeDelta",
+     [NSNumber numberWithDouble:_region.span.longitudeDelta],@"longitudeDelta",
+     NUMBOOL(animated),@"animated",nil];
+    
+    [self.proxy fireEvent:event withObject:object];
 }
 
 - (void)fireClickEvent:(MKAnnotationView *) pinview source:(NSString *)source
@@ -1165,12 +1177,12 @@
 		title = [NSNull null];
 	}
 
-	NSNumber * indexNumber = [pinview tag];
+	NSInteger indexNumber = [pinview tag];
 	id clicksource = source ? source : (id)[NSNull null];
 	
 	NSDictionary * event = [NSDictionary dictionaryWithObjectsAndKeys:
 			clicksource,@"clicksource",	viewProxy,@"annotation", mapProxy,@"map",
-			title,@"title",	indexNumber,@"index", nil];
+			title,@"title",	NUMINTEGER(indexNumber),@"index", nil];
 
     [self doClickEvent:viewProxy mapProxy:mapProxy event:event];
 }